+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 - Added the Lunchbox allowing convenient carrying of food, as well as eating directly out of the item
 - Added the Mob Net to carry small mobs
+
+### Changes
+
+- Inventory interaction tooltip only shows on container items which have the feature enabled in config